server/pyproject.toml
[project]
    dependencies    = ["asyncio >=3.4.3", "pillow >= 11.1.0", "websockets>=15.0"]
    description     = "SpectaclesXR Teleop Server"
    name            = "teleop"
    readme          = "README.md"
    requires-python = ">=3.12"
    version         = "0.1.0"
